switching test_qa_metrics and test_qa_tests
